Big refactor to rely on Mongo4Cats
This way I don't have to maintain so much of the app
