+++ core/src/main/scala/zongo/MongoDoc.scala
@@ -1,0 +1,7 @@
+package zongo
+
+import mongo4cats.bson._
+
+abstract class MongoDoc extends Product {
+  def _id: Option[MongoId]
+}
