+++ core/src/test/resources/application.conf
