+++ core/src/test/scala/zongo/support/package.scala
@@ -1,0 +1,9 @@
+package zongo
+
+// import mongo4cats.bson._
+
+import zio.Has
+
+package object support {
+  // type ItemsRepo = Has[ItemsRepo.Service]
+}
