+++ project/Dependencies.scala
@@ -3,15 +3,15 @@ import sbt._
 object Dependencies {
   object V {
     val betterMonadicFor = "0.3.1"
-    val mongo            = "4.2.3"
+    val mongo4cats       = "0.4.0"
     val circe            = "0.14.1"
-    val zio              = "1.0.10"
-    val zioCats          = "2.4.0.0"
-    val zioConfig        = "1.0.4"
+    val zio              = "1.0.12"
+    val zioCats          = "3.1.1.0"
+    val zioConfig        = "1.0.6"
     val zioLogging       = "0.5.8"
-    val zioMagic         = "0.2.6"
-    val zioReactStreams  = "1.3.4"
+    val zioMagic         = "0.3.6"
     val zioPrelude       = "1.0.0-RC6"
+    val zioReactStreams  = "1.3.4"
   }
 
   object CompilerPlugins {
@@ -23,16 +23,18 @@ object Dependencies {
   /** Used in RepoTests */
   object Circe {
     val libs = Seq(
-      "io.circe" %% "circe-core"    % V.circe,
-      "io.circe" %% "circe-generic" % V.circe,
-      "io.circe" %% "circe-parser"  % V.circe
+      "io.circe"           %% "circe-core"       % V.circe,
+      "io.circe"           %% "circe-generic"    % V.circe,
+      "io.circe"           %% "circe-parser"     % V.circe,
+      "io.github.kirill5k" %% "mongo4cats-circe" % V.mongo4cats
     )
   }
 
   object MongoDB {
     val libs = Seq(
-      "org.mongodb.scala" %% "mongo-scala-bson"   % V.mongo,
-      "org.mongodb.scala" %% "mongo-scala-driver" % V.mongo
+      "io.github.kirill5k" %% "mongo4cats-core" % V.mongo4cats
+      // "org.mongodb.scala" %% "mongo-scala-bson"   % V.mongo,
+      // "org.mongodb.scala" %% "mongo-scala-driver" % V.mongo
     )
   }
 
@@ -60,4 +62,10 @@ object Dependencies {
       "io.github.kitlangton" %% "zio-magic"                   % V.zioMagic
     )
   }
+
+  def core =
+    MongoDB.libs ++ Testing.libs ++ ZIO.libs
+
+  def circe =
+    Circe.libs ++ Testing.libs
 }
