--- src/main/scala/zongo/MongoDoc.scala
@@ -1,9 +1,0 @@
-package zongo
-
-import org.mongodb.scala.bson._
-import org.mongodb.scala.bson.codecs.Macros
-import org.bson.codecs.configuration.CodecProvider
-
-abstract class MongoDoc extends Product {
-  def _id: Option[MongoId]
-}
